split infrastructure layer into adapters and entry-points
