+++ src/infrastructure/adapters/db/models/index.js
@@ -5,7 +5,7 @@ const path = require('path');
 const Sequelize = require('sequelize');
 const process = require('process');
 const basename = path.basename(__filename);
-const config = require('../../../common/config');
+const config = require('../../../../common/config');
 const db = {};
 
 let sequelize;
