+++ src/infrastructure/adapters/providers/email-providers/emailProviderManager.js
@@ -1,5 +1,5 @@
-const IEmailProvider = require('../../../domain/interfaces/providers/emailProviderInterface');
-const Result = require('../../../domain/types/result');
+const IEmailProvider = require('../../../../domain/interfaces/providers/emailProviderInterface');
+const Result = require('../../../../domain/types/result');
 // const { logProviderResponse } = require('../../utils/logger');
 // const path = require('path');
