+++ src/infrastructure/entry-points/cron/main.js
