+++ test/unit/providers/weatherProviderManager.test.js
@@ -1,5 +1,5 @@
 const { expect } = require('chai');
-const WeatherProviderManager = require('../../../src/infrastructure/providers/weather-providers/weatherProviderManager');
+const WeatherProviderManager = require('../../../src/infrastructure/adapters/providers/weather-providers/weatherProviderManager');
 const { WeatherProviderMock1, WeatherProviderMock2 } = require('../../mocks/providers/weatherProviders.mock');
 const LoggerMock = require('../../mocks/utils/logger.mock');
